More experiments with transparent background.
As hiding and showing the window is far to slow and also OpenGL
overlays are only supported on few GPUs a different approach is
used.
By enabling a transparent background it is now possible to clear
the background with OpenGL, take a screenshot and show the links
again. This should be sufficiently fast and also work on every
platform with a decent GPU.